Refactor scope list to match ml_config.json and remove unused /usr/local/lib/python
